Resolved several issues relating to cursor actions in ThreeDWorld by deriving an anonymous MouseInput and flipping the y coordinate of the AWT mouse events (as suggested on JME forums). Now using AWT mouse position directly in context menu because this needs to be "unflipped".
Added toolbar buttons for control/release agent (not finalized).

Added translate to cursor on "Add Entity" and "Add Agent", so that a right
click add will place the entity at the click position. Toolbar and menu
adds may need a special case (0, 0, 0)?

Reorganized files to try to simplify structure.

Eliminated ActionManager as a class in favor of a map of actions since
it was only acting as a container anyway.

Added an id counter to ThreeDWorld to provide unique names for both agents
and entities. This should fix various selection issues.

Added a camera home action and placed it on the toolbar.

Fixed errors in Run and Pause actions from copying the Update action.

Fixed offset bounding box problems in the SelectionController, where
the center of the bounding box was not correctly defined. The mouse is
now properly placed on surfaces and its bounding box encloses it.

Copied AwtPanelsContext and AwtPanel to prepare for rewrite (support agent
cameras).

Fixed problems with Entity set transforms, where kinematic entities
(Agent is kinematic to allow direct movement) could not be transformed.