Implemented grayscale and threshold (black & white) filters for views.
Refactored ThreeDWorld frame and panel resize implementation so that the
main view (spectator) will always be resized to fit the panel, but agent
views will only be scaled to fit the panel.

Implemented synchronized update mode for ThreeDWorld (uncheck real time
in preferences) and fixed a bug with the Step update where objects would
shake after a step.

Fixed a bug in EditWorldPreferences where the engine was not being
unpaused. Most preferences are not currently being applied and this dialog
is still very bad.

Fixed a bug where animations would continue during pause.